Duration Option
I've added an option where you require a fixed duration for the animation. Let's say 1 second. So it calculates the speed of the animation based on the number of animated characters to make sure each transition takes the same length of time.

I'm about to fork this code quite dramatically - so I thought I'd commit this before I make it completely unusable!